Add instructions for poetry keyring check bug 1917
See: https://github.com/python-poetry/poetry/issues/1917
Prior duplicate with exact matching error message:
   https://github.com/python-poetry/poetry/issues/7032

Encountered a poetry bug: "Failed to unlock the collection"
Unsure if this bug is per-project. If it recurs:
   Temporary fix:
   ```PYTHON_KEYRING_BACKEND=keyring.backends.null.Keyring```
      Or switch to a working backend.
   Permanent fix (disables Python keyring facilities):
   ```python3 -m keyring --disable```

I am uncertain if this bug is elicited by particular project settings,
but it is by particular environments.